+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.2.3
 
 - add phfit for group data
+- add phfit for grouptrunc and trunc poi
 
 # 0.2.2